Change auto gen and type id to telegram chat id type int

+++ src/main/java/com/foxminded/telebot/repository/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserRepository extends JpaRepository<TelegramUser, Long> {
+public interface UserRepository extends JpaRepository<TelegramUser, Integer> {
 }
